Extract grants validations and fix them
