+++ pkg/sdk/grants.go
@@ -2,17 +2,8 @@ package sdk
 
 import (
 	"context"
-	"fmt"
 	"strings"
 	"time"
-)
-
-var (
-	_ validatable = new(GrantPrivilegesToAccountRoleOptions)
-	_ validatable = new(RevokePrivilegesFromAccountRoleOptions)
-	_ validatable = new(grantPrivilegeToShareOptions)
-	_ validatable = new(revokePrivilegeFromShareOptions)
-	_ validatable = new(ShowGrantOptions)
 )
 
 type Grants interface {
@@ -95,22 +86,6 @@ type GrantPrivilegesToAccountRoleOptions struct {
 	WithGrantOption *bool                       `ddl:"keyword" sql:"WITH GRANT OPTION"`
 }
 
-func (opts *GrantPrivilegesToAccountRoleOptions) validate() error {
-	if !valueSet(opts.privileges) {
-		return fmt.Errorf("privileges must be set")
-	}
-	if err := opts.privileges.validate(); err != nil {
-		return err
-	}
-	if !valueSet(opts.on) {
-		return fmt.Errorf("on must be set")
-	}
-	if err := opts.on.validate(); err != nil {
-		return err
-	}
-	return nil
-}
-
 type AccountRoleGrantPrivileges struct {
 	GlobalPrivileges        []GlobalPrivilege        `ddl:"-"`
 	AccountObjectPrivileges []AccountObjectPrivilege `ddl:"-"`
@@ -119,40 +94,11 @@ type AccountRoleGrantPrivileges struct {
 	AllPrivileges           *bool                    `ddl:"keyword" sql:"ALL PRIVILEGES"`
 }
 
-func (v *AccountRoleGrantPrivileges) validate() error {
-	if !exactlyOneValueSet(v.AllPrivileges, v.GlobalPrivileges, v.AccountObjectPrivileges, v.SchemaPrivileges, v.SchemaObjectPrivileges) {
-		return fmt.Errorf("exactly one of AllPrivileges, GlobalPrivileges, AccountObjectPrivileges, SchemaPrivileges, or SchemaObjectPrivileges must be set")
-	}
-	return nil
-}
-
 type AccountRoleGrantOn struct {
 	Account       *bool                 `ddl:"keyword" sql:"ACCOUNT"`
 	AccountObject *GrantOnAccountObject `ddl:"-"`
 	Schema        *GrantOnSchema        `ddl:"-"`
 	SchemaObject  *GrantOnSchemaObject  `ddl:"-"`
-}
-
-func (v *AccountRoleGrantOn) validate() error {
-	if !exactlyOneValueSet(v.Account, v.AccountObject, v.Schema, v.SchemaObject) {
-		return fmt.Errorf("exactly one of Account, AccountObject, Schema, or SchemaObject must be set")
-	}
-	if valueSet(v.AccountObject) {
-		if err := v.AccountObject.validate(); err != nil {
-			return err
-		}
-	}
-	if valueSet(v.Schema) {
-		if err := v.Schema.validate(); err != nil {
-			return err
-		}
-	}
-	if valueSet(v.SchemaObject) {
-		if err := v.SchemaObject.validate(); err != nil {
-			return err
-		}
-	}
-	return nil
 }
 
 type GrantOnAccountObject struct {
@@ -165,24 +111,10 @@ type GrantOnAccountObject struct {
 	ReplicationGroup *AccountObjectIdentifier `ddl:"identifier" sql:"REPLICATION GROUP"`
 }
 
-func (v *GrantOnAccountObject) validate() error {
-	if !exactlyOneValueSet(v.User, v.ResourceMonitor, v.Warehouse, v.Database, v.Integration, v.FailoverGroup, v.ReplicationGroup) {
-		return fmt.Errorf("exactly one of User, ResourceMonitor, Warehouse, Database, Integration, FailoverGroup, or ReplicationGroup must be set")
-	}
-	return nil
-}
-
 type GrantOnSchema struct {
 	Schema                  *DatabaseObjectIdentifier `ddl:"identifier" sql:"SCHEMA"`
 	AllSchemasInDatabase    *AccountObjectIdentifier  `ddl:"identifier" sql:"ALL SCHEMAS IN DATABASE"`
 	FutureSchemasInDatabase *AccountObjectIdentifier  `ddl:"identifier" sql:"FUTURE SCHEMAS IN DATABASE"`
-}
-
-func (v *GrantOnSchema) validate() error {
-	if !exactlyOneValueSet(v.Schema, v.AllSchemasInDatabase, v.FutureSchemasInDatabase) {
-		return fmt.Errorf("exactly one of Schema, AllSchemasInDatabase, or FutureSchemasInDatabase must be set")
-	}
-	return nil
 }
 
 type GrantOnSchemaObject struct {
@@ -191,24 +123,10 @@ type GrantOnSchemaObject struct {
 	Future       *GrantOnSchemaObjectIn `ddl:"keyword" sql:"FUTURE"`
 }
 
-func (v *GrantOnSchemaObject) validate() error {
-	if !exactlyOneValueSet(v.SchemaObject, v.All, v.Future) {
-		return fmt.Errorf("exactly one of Object, AllIn or Future must be set")
-	}
-	return nil
-}
-
 type GrantOnSchemaObjectIn struct {
 	PluralObjectType PluralObjectType          `ddl:"keyword" sql:"ALL"`
 	InDatabase       *AccountObjectIdentifier  `ddl:"identifier" sql:"IN DATABASE"`
 	InSchema         *DatabaseObjectIdentifier `ddl:"identifier" sql:"IN SCHEMA"`
-}
-
-func (v *GrantOnSchemaObjectIn) validate() error {
-	if !exactlyOneValueSet(v.PluralObjectType, v.InDatabase, v.InSchema) {
-		return fmt.Errorf("exactly one of PluralObjectType, InDatabase, or InSchema must be set")
-	}
-	return nil
 }
 
 func (v *grants) GrantPrivilegesToAccountRole(ctx context.Context, privileges *AccountRoleGrantPrivileges, on *AccountRoleGrantOn, role AccountObjectIdentifier, opts *GrantPrivilegesToAccountRoleOptions) error {
@@ -242,28 +160,6 @@ type RevokePrivilegesFromAccountRoleOptions struct {
 	Cascade        *bool                       `ddl:"keyword" sql:"CASCADE"`
 }
 
-func (opts *RevokePrivilegesFromAccountRoleOptions) validate() error {
-	if !valueSet(opts.privileges) {
-		return fmt.Errorf("privileges must be set")
-	}
-	if err := opts.privileges.validate(); err != nil {
-		return err
-	}
-	if !valueSet(opts.on) {
-		return fmt.Errorf("on must be set")
-	}
-	if err := opts.on.validate(); err != nil {
-		return err
-	}
-	if !validObjectidentifier(opts.accountRole) {
-		return ErrInvalidObjectIdentifier
-	}
-	if everyValueSet(opts.Restrict, opts.Cascade) {
-		return fmt.Errorf("either Restrict or Cascade can be set, or neither but not both")
-	}
-	return nil
-}
-
 func (v *grants) RevokePrivilegesFromAccountRole(ctx context.Context, privileges *AccountRoleGrantPrivileges, on *AccountRoleGrantOn, role AccountObjectIdentifier, opts *RevokePrivilegesFromAccountRoleOptions) error {
 	if opts == nil {
 		opts = &RevokePrivilegesFromAccountRoleOptions{}
@@ -292,19 +188,6 @@ type grantPrivilegeToShareOptions struct {
 	to        AccountObjectIdentifier  `ddl:"identifier" sql:"TO SHARE"`
 }
 
-func (opts *grantPrivilegeToShareOptions) validate() error {
-	if !validObjectidentifier(opts.to) {
-		return ErrInvalidObjectIdentifier
-	}
-	if !valueSet(opts.On) || opts.privilege == "" {
-		return fmt.Errorf("on and privilege are required")
-	}
-	if !exactlyOneValueSet(opts.On.Database, opts.On.Schema, opts.On.Function, opts.On.Table, opts.On.View) {
-		return fmt.Errorf("only one of database, schema, function, table, or view can be set")
-	}
-	return nil
-}
-
 type GrantPrivilegeToShareOn struct {
 	Database AccountObjectIdentifier  `ddl:"identifier" sql:"DATABASE"`
 	Schema   DatabaseObjectIdentifier `ddl:"identifier" sql:"SCHEMA"`
@@ -313,28 +196,9 @@ type GrantPrivilegeToShareOn struct {
 	View     SchemaObjectIdentifier   `ddl:"identifier" sql:"VIEW"`
 }
 
-func (v *GrantPrivilegeToShareOn) validate() error {
-	if !exactlyOneValueSet(v.Database, v.Schema, v.Function, v.Table, v.View) {
-		return fmt.Errorf("only one of database, schema, function, table, or view can be set")
-	}
-	if valueSet(v.Table) {
-		if err := v.Table.validate(); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
 type OnTable struct {
 	Name        SchemaObjectIdentifier   `ddl:"identifier" sql:"TABLE"`
 	AllInSchema DatabaseObjectIdentifier `ddl:"identifier" sql:"ALL TABLES IN SCHEMA"`
-}
-
-func (v *OnTable) validate() error {
-	if !exactlyOneValueSet(v.Name, v.AllInSchema) {
-		return fmt.Errorf("only one of name or allInSchema can be set")
-	}
-	return nil
 }
 
 func (v *grants) GrantPrivilegeToShare(ctx context.Context, privilege ObjectPrivilege, on *GrantPrivilegeToShareOn, to AccountObjectIdentifier) error {
@@ -361,24 +225,6 @@ type revokePrivilegeFromShareOptions struct {
 	from      AccountObjectIdentifier     `ddl:"identifier" sql:"FROM SHARE"`
 }
 
-func (opts *revokePrivilegeFromShareOptions) validate() error {
-	if !validObjectidentifier(opts.from) {
-		return ErrInvalidObjectIdentifier
-	}
-	if !valueSet(opts.On) || opts.privilege == "" {
-		return fmt.Errorf("on and privilege are required")
-	}
-	if !exactlyOneValueSet(opts.On.Database, opts.On.Schema, opts.On.Table, opts.On.View) {
-		return fmt.Errorf("only one of database, schema, function, table, or view can be set")
-	}
-
-	if err := opts.On.validate(); err != nil {
-		return err
-	}
-
-	return nil
-}
-
 type RevokePrivilegeFromShareOn struct {
 	Database AccountObjectIdentifier  `ddl:"identifier" sql:"DATABASE"`
 	Schema   DatabaseObjectIdentifier `ddl:"identifier" sql:"SCHEMA"`
@@ -386,29 +232,9 @@ type RevokePrivilegeFromShareOn struct {
 	View     *OnView                  `ddl:"-"`
 }
 
-func (v *RevokePrivilegeFromShareOn) validate() error {
-	if !exactlyOneValueSet(v.Database, v.Schema, v.Table, v.View) {
-		return fmt.Errorf("only one of database, schema, table, or view can be set")
-	}
-	if valueSet(v.Table) {
-		return v.Table.validate()
-	}
-	if valueSet(v.View) {
-		return v.View.validate()
-	}
-	return nil
-}
-
 type OnView struct {
 	Name        SchemaObjectIdentifier   `ddl:"identifier" sql:"VIEW"`
 	AllInSchema DatabaseObjectIdentifier `ddl:"identifier" sql:"ALL VIEWS IN SCHEMA"`
-}
-
-func (v *OnView) validate() error {
-	if !exactlyOneValueSet(v.Name, v.AllInSchema) {
-		return fmt.Errorf("only one of name or allInSchema can be set")
-	}
-	return nil
 }
 
 func (v *grants) RevokePrivilegeFromShare(ctx context.Context, privilege ObjectPrivilege, on *RevokePrivilegeFromShareOn, id AccountObjectIdentifier) error {
@@ -436,16 +262,6 @@ type ShowGrantOptions struct {
 	To     *ShowGrantsTo `ddl:"keyword" sql:"TO"`
 	Of     *ShowGrantsOf `ddl:"keyword" sql:"OF"`
 	In     *ShowGrantsIn `ddl:"keyword" sql:"IN"`
-}
-
-func (opts *ShowGrantOptions) validate() error {
-	if everyValueNil(opts.On, opts.To, opts.Of, opts.In) {
-		return fmt.Errorf("at least one of on, to, of, or in is required")
-	}
-	if !exactlyOneValueSet(opts.On, opts.To, opts.Of, opts.In) {
-		return fmt.Errorf("only one of on, to, of, or in can be set")
-	}
-	return nil
 }
 
 type ShowGrantsIn struct {
