+++ pkg/sdk/grants_validations.go
@@ -1,0 +1,198 @@
+package sdk
+
+import (
+	"fmt"
+)
+
+var (
+	_ validatable = new(GrantPrivilegesToAccountRoleOptions)
+	_ validatable = new(RevokePrivilegesFromAccountRoleOptions)
+	_ validatable = new(grantPrivilegeToShareOptions)
+	_ validatable = new(revokePrivilegeFromShareOptions)
+	_ validatable = new(ShowGrantOptions)
+)
+
+func (opts *GrantPrivilegesToAccountRoleOptions) validate() error {
+	if !valueSet(opts.privileges) {
+		return fmt.Errorf("privileges must be set")
+	}
+	if err := opts.privileges.validate(); err != nil {
+		return err
+	}
+	if !valueSet(opts.on) {
+		return fmt.Errorf("on must be set")
+	}
+	if err := opts.on.validate(); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (v *AccountRoleGrantPrivileges) validate() error {
+	if !exactlyOneValueSet(v.AllPrivileges, v.GlobalPrivileges, v.AccountObjectPrivileges, v.SchemaPrivileges, v.SchemaObjectPrivileges) {
+		return fmt.Errorf("exactly one of AllPrivileges, GlobalPrivileges, AccountObjectPrivileges, SchemaPrivileges, or SchemaObjectPrivileges must be set")
+	}
+	return nil
+}
+
+func (v *AccountRoleGrantOn) validate() error {
+	if !exactlyOneValueSet(v.Account, v.AccountObject, v.Schema, v.SchemaObject) {
+		return fmt.Errorf("exactly one of Account, AccountObject, Schema, or SchemaObject must be set")
+	}
+	if valueSet(v.AccountObject) {
+		if err := v.AccountObject.validate(); err != nil {
+			return err
+		}
+	}
+	if valueSet(v.Schema) {
+		if err := v.Schema.validate(); err != nil {
+			return err
+		}
+	}
+	if valueSet(v.SchemaObject) {
+		if err := v.SchemaObject.validate(); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (v *GrantOnAccountObject) validate() error {
+	if !exactlyOneValueSet(v.User, v.ResourceMonitor, v.Warehouse, v.Database, v.Integration, v.FailoverGroup, v.ReplicationGroup) {
+		return fmt.Errorf("exactly one of User, ResourceMonitor, Warehouse, Database, Integration, FailoverGroup, or ReplicationGroup must be set")
+	}
+	return nil
+}
+
+func (v *GrantOnSchema) validate() error {
+	if !exactlyOneValueSet(v.Schema, v.AllSchemasInDatabase, v.FutureSchemasInDatabase) {
+		return fmt.Errorf("exactly one of Schema, AllSchemasInDatabase, or FutureSchemasInDatabase must be set")
+	}
+	return nil
+}
+
+func (v *GrantOnSchemaObject) validate() error {
+	if !exactlyOneValueSet(v.SchemaObject, v.All, v.Future) {
+		return fmt.Errorf("exactly one of Object, AllIn or Future must be set")
+	}
+	if valueSet(v.All) {
+		if err := v.All.validate(); err != nil {
+			return err
+		}
+	}
+	if valueSet(v.Future) {
+		if err := v.Future.validate(); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (v *GrantOnSchemaObjectIn) validate() error {
+	if !exactlyOneValueSet(v.PluralObjectType, v.InDatabase, v.InSchema) {
+		return fmt.Errorf("exactly one of PluralObjectType, InDatabase, or InSchema must be set")
+	}
+	return nil
+}
+
+func (opts *RevokePrivilegesFromAccountRoleOptions) validate() error {
+	if !valueSet(opts.privileges) {
+		return fmt.Errorf("privileges must be set")
+	}
+	if err := opts.privileges.validate(); err != nil {
+		return err
+	}
+	if !valueSet(opts.on) {
+		return fmt.Errorf("on must be set")
+	}
+	if err := opts.on.validate(); err != nil {
+		return err
+	}
+	if !validObjectidentifier(opts.accountRole) {
+		return ErrInvalidObjectIdentifier
+	}
+	if everyValueSet(opts.Restrict, opts.Cascade) {
+		return fmt.Errorf("either Restrict or Cascade can be set, or neither but not both")
+	}
+	return nil
+}
+
+func (opts *grantPrivilegeToShareOptions) validate() error {
+	if !validObjectidentifier(opts.to) {
+		return ErrInvalidObjectIdentifier
+	}
+	if !valueSet(opts.On) || opts.privilege == "" {
+		return fmt.Errorf("on and privilege are required")
+	}
+	if err := opts.On.validate(); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (v *GrantPrivilegeToShareOn) validate() error {
+	if !exactlyOneValueSet(v.Database, v.Schema, v.Function, v.Table, v.View) {
+		return fmt.Errorf("only one of database, schema, function, table, or view can be set")
+	}
+	if valueSet(v.Table) {
+		if err := v.Table.validate(); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (v *OnTable) validate() error {
+	if !exactlyOneValueSet(v.Name, v.AllInSchema) {
+		return fmt.Errorf("only one of name or allInSchema can be set")
+	}
+	return nil
+}
+
+func (opts *revokePrivilegeFromShareOptions) validate() error {
+	if !validObjectidentifier(opts.from) {
+		return ErrInvalidObjectIdentifier
+	}
+	if !valueSet(opts.On) || opts.privilege == "" {
+		return fmt.Errorf("on and privilege are required")
+	}
+	if !exactlyOneValueSet(opts.On.Database, opts.On.Schema, opts.On.Table, opts.On.View) {
+		return fmt.Errorf("only one of database, schema, function, table, or view can be set")
+	}
+
+	if err := opts.On.validate(); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (v *RevokePrivilegeFromShareOn) validate() error {
+	if !exactlyOneValueSet(v.Database, v.Schema, v.Table, v.View) {
+		return fmt.Errorf("only one of database, schema, table, or view can be set")
+	}
+	if valueSet(v.Table) {
+		return v.Table.validate()
+	}
+	if valueSet(v.View) {
+		return v.View.validate()
+	}
+	return nil
+}
+
+func (v *OnView) validate() error {
+	if !exactlyOneValueSet(v.Name, v.AllInSchema) {
+		return fmt.Errorf("only one of name or allInSchema can be set")
+	}
+	return nil
+}
+
+func (opts *ShowGrantOptions) validate() error {
+	if everyValueNil(opts.On, opts.To, opts.Of, opts.In) {
+		return fmt.Errorf("at least one of on, to, of, or in is required")
+	}
+	if !exactlyOneValueSet(opts.On, opts.To, opts.Of, opts.In) {
+		return fmt.Errorf("only one of on, to, of, or in can be set")
+	}
+	return nil
+}
